utils: move dl to utils::http
