+++ src/crates.rs
@@ -1,5 +1,4 @@
 use dirs::{CRATES_DIR, GH_MIRRORS_DIR};
-use dl;
 use errors::*;
 use flate2::read::GzDecoder;
 use std::fmt;
@@ -164,7 +163,7 @@ fn dl_registry(name: &str, vers: &str, dir: &Path) -> Result<()> {
     }
     info!("downloading crate {}-{} to {}", name, vers, dir.display());
     let url = format!("{0}/{1}/{1}-{2}.crate", CRATES_ROOT, name, vers);
-    let bin = dl::download(&url).chain_err(|| format!("unable to download {}", url))?;
+    let bin = ::utils::http::get(&url).chain_err(|| format!("unable to download {}", url))?;
 
     fs::create_dir_all(&dir)?;
